Refactor translation files from TS to JSON
Refactor the translation files from TypeScript to JSON and update the code to use the new JSON files.

* **Translation Files:**
  - Rename `src/locales/language/en.ts` to `src/locales/language/en.json`.
  - Rename `src/locales/language/fr.ts` to `src/locales/language/fr.json`.
  - Convert the TypeScript objects to JSON format.
  - Remove duplicate keys in the JSON files.

* **Code Updates:**
  - Update import statements in various command and event files to import translations from JSON files.
  - Update the usage of translation keys to reflect the new JSON structure.

* **Package Management:**
  - Add `package-lock.json` to manage dependencies.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/Mara-Li/tickette-bot?shareId=XXXX-XXXX-XXXX-XXXX).
