+++ src/commands/editEmbed.ts
@@ -12,8 +12,8 @@ import {
 } from "discord.js";
 
 import { ln } from "../locales";
-import en from "../locales/language/en";
-import fr from "../locales/language/fr";
+import en from "../locales/language/en.json";
+import fr from "../locales/language/fr.json";
 import { createEmbed } from "../tickets/modals";
 import { downloadJSONTemplate, parseLinkFromDiscord } from "../tickets/template";
