+++ src/commands/ping.ts
@@ -1,7 +1,7 @@
 import * as Djs from "discord.js";
 import { ln } from "../locales";
-import en from "../locales/language/en";
-import fr from "../locales/language/fr";
+import en from "../locales/language/en.json";
+import fr from "../locales/language/fr.json";
 import { createFile, downloadJSONTemplate } from "../tickets/template";
 
 export const pingRole = {
