+++ src/commands/rename.ts
@@ -6,8 +6,8 @@ import {
 } from "discord.js";
 
 import { ln } from "../locales";
-import en from "../locales/language/en";
-import fr from "../locales/language/fr";
+import en from "../locales/language/en.json";
+import fr from "../locales/language/fr.json";
 import { renameThread } from "../tickets/editname";
 import { downloadJSONTemplate } from "../tickets/template";
