+++ src/commands/role.ts
@@ -6,8 +6,8 @@ import {
 } from "discord.js";
 
 import { ln } from "../locales";
-import en from "../locales/language/en";
-import fr from "../locales/language/fr";
+import en from "../locales/language/en.json";
+import fr from "../locales/language/fr.json";
 import { addRoleToTemplate, removeRoleToTemplate } from "../tickets/editRole";
 import { downloadJSONTemplate } from "../tickets/template";
